refactor: dto 코드 경로 분리 및 ID값 type 변경
