+++ src/vote/dto/vote.response.dto.ts
